refactor overview code
